# Measurement after optimization