--- a/benchmarks/definitions/curated/06-cloud-flare-redos.toml
+++ b/benchmarks/definitions/curated/06-cloud-flare-redos.toml
@@ -108,6 +108,7 @@ engines = [
   'icu',
   'java/hotspot',
   'javascript/v8',
+  'lean',
   'pcre2',
   'pcre2/jit',
   'perl',