--- a/benchmarks/definitions/curated/11-unstructured-to-json.toml
+++ b/benchmarks/definitions/curated/11-unstructured-to-json.toml
@@ -62,6 +62,8 @@ engines = [
   'icu',
   'java/hotspot',
   'javascript/v8',
+  # Failed to parse regex
+  # 'lean',
   'pcre2',
   'pcre2/jit',
   'perl',
@@ -89,6 +91,8 @@ engines = [
   'dotnet/nobacktrack',
   'go/regexp',
   'icu',
+  # Failed to parse regex
+  # 'lean',
   'pcre2',
   'pcre2/jit',
   'python/re',