--- a/benchmarks/definitions/test/dot.toml
+++ b/benchmarks/definitions/test/dot.toml
@@ -21,7 +21,7 @@ engines = [
   'icu',
   'java/hotspot',
   'javascript/v8',
-  # 'lean',
+  'lean',
   'pcre2',
   'pcre2/jit',
   'perl',