--- a/benchmarks/definitions/unicode/codepoints.toml
+++ b/benchmarks/definitions/unicode/codepoints.toml
@@ -170,7 +170,6 @@ engines = [
   'pcre2/jit',
   'python/regex',
   're2',
-  'regress',
   'rust/regex',
   'rust/regexold',
 ]