--- a/benchmarks/definitions/unicode/codepoints.toml
+++ b/benchmarks/definitions/unicode/codepoints.toml
@@ -55,7 +55,9 @@ haystack = { path = "unicode/allcodepoints.txt" }
 unicode = true
 count = [
   # Very likely due to update to Unicode 15.1 (a month old at time of writing).
-  { engine = '^(python/regex|re2)$', count = 1250 },
+  { engine = '^re2$', count = 1250 },
+  # Very likely due to update to Unicode 16 (a month old at time of writing).
+  { engine = '^python/regex$', count = 1289 },
   # Maybe a different version of Unicode?
   { engine = 'pcre2/jit', count = 1203 },
   { engine = '.*', count = 1244 },
@@ -100,7 +102,11 @@ haystack = { path = "unicode/allcodepoints.txt" }
 unicode = true
 count = [
   # Very likely due to update to Unicode 15.1 (a month old at time of writing).
-  { engine = 'python/regex', count = 1250 },
+  { engine = '^re2$', count = 1250 },
+  # Very likely due to update to Unicode 16 (a month old at time of writing).
+  { engine = '^python/regex$', count = 1289 },
+  # Maybe a different version of Unicode?
+  { engine = 'pcre2/jit', count = 1203 },
   { engine = '.*', count = 1244 },
 ]
 engines = [