--- a/benchmarks/definitions/unicode/codepoints.toml
+++ b/benchmarks/definitions/unicode/codepoints.toml
@@ -57,7 +57,7 @@ count = [
   # Very likely due to update to Unicode 15.1 (a month old at time of writing).
   { engine = '^re2$', count = 1250 },
   # Very likely due to update to Unicode 16 (a month old at time of writing).
-  { engine = '^python/regex$', count = 1289 },
+  { engine = '^(python/regex|rust/regex)$', count = 1289 },
   # Maybe a different version of Unicode?
   { engine = 'pcre2/jit', count = 1203 },
   { engine = '.*', count = 1244 },
@@ -104,7 +104,7 @@ count = [
   # Very likely due to update to Unicode 15.1 (a month old at time of writing).
   { engine = '^re2$', count = 1250 },
   # Very likely due to update to Unicode 16 (a month old at time of writing).
-  { engine = '^python/regex$', count = 1289 },
+  { engine = '^(python/regex|rust/regex)$', count = 1289 },
   # Maybe a different version of Unicode?
   { engine = 'pcre2/jit', count = 1203 },
   { engine = '.*', count = 1244 },