--- a/engines/rust/aho-corasick/Cargo.toml
+++ b/engines/rust/aho-corasick/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "main"
-version = "1.1.2"
+version = "1.1.3"
 edition = "2021"
 
 [[bin]]
@@ -21,7 +21,7 @@ lexopt = "0.3.0"
 # parsing the Cargo.lock file would be inappropriate as that doesn't mean the
 # engine is available to run. Otherwise, it would be fine to just rely on
 # 'Cargo.lock' for version pinning and we could drop the '=' constraint here.
-aho-corasick = "=1.1.2"
+aho-corasick = "=1.1.3"
 
 [dependencies.klv]
 path = "../../../shared/klv"