--- a/engines/rust/memchr/Cargo.toml
+++ b/engines/rust/memchr/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rust-memchr-memmem"
-version = "2.6.4"
+version = "2.7.4"
 edition = "2021"
 
 [[bin]]
@@ -21,7 +21,7 @@ lexopt = "0.3.0"
 # parsing the Cargo.lock file would be inappropriate as that doesn't mean the
 # engine is available to run. Otherwise, it would be fine to just rely on
 # 'Cargo.lock' for version pinning and we could drop the '=' constraint here.
-memchr = "=2.6.4"
+memchr = "=2.7.4"
 
 [dependencies.klv]
 path = "../../../shared/klv"