--- a/engines/rust/regex-lite/Cargo.toml
+++ b/engines/rust/regex-lite/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "main"
-version = "0.1.5"
+version = "0.1.6"
 edition = "2021"
 
 [[bin]]
@@ -21,7 +21,7 @@ lexopt = "0.3.0"
 # parsing the Cargo.lock file would be inappropriate as that doesn't mean the
 # engine is available to run. Otherwise, it would be fine to just rely on
 # 'Cargo.lock' for version pinning and we could drop the '=' constraint here.
-regex-lite = { version = "=0.1.5" }
+regex-lite = { version = "=0.1.6" }
 
 [dependencies.klv]
 path = "../../../shared/klv"