--- a/engines/rust/regex-syntax/Cargo.toml
+++ b/engines/rust/regex-syntax/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "main"
-version = "0.8.2"
+version = "0.8.5"
 edition = "2021"
 publish = false
 
@@ -13,7 +13,7 @@ anyhow = "1.0.69"
 lexopt = "0.3.0"
 # We use this just for turning ASTs and HIRs into regex matching machines in
 # order to perform the verification step after "compiling" to an AST or HIR.
-regex-automata = { version = "0.4.3", features = ["logging"] }
+regex-automata = { version = "0.4.8", features = ["logging"] }
 # We use an '=' here to force a specific version. It is okay to update it,
 # but if you do, change the version of this Cargo package as well. This way,
 # the version of the runner matches the version of engine, which makes it
@@ -24,7 +24,7 @@ regex-automata = { version = "0.4.3", features = ["logging"] }
 # parsing the Cargo.lock file would be inappropriate as that doesn't mean the
 # engine is available to run. Otherwise, it would be fine to just rely on
 # 'Cargo.lock' for version pinning and we could drop the '=' constraint here.
-regex-syntax = "=0.8.2"
+regex-syntax = "=0.8.5"
 
 [dependencies.klv]
 path = "../../../shared/klv"