--- a/engines/rust/regex/Cargo.toml
+++ b/engines/rust/regex/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "main"
-version = "1.10.2"
+version = "1.11.0"
 edition = "2021"
 
 [[bin]]
@@ -24,11 +24,11 @@ lexopt = "0.3.0"
 # parsing the Cargo.lock file would be inappropriate as that doesn't mean the
 # engine is available to run. Otherwise, it would be fine to just rely on
 # 'Cargo.lock' for version pinning and we could drop the '=' constraint here.
-regex = { version = "=1.10.2", features = ["logging", "perf-dfa-full"] }
+regex = { version = "=1.11.0", features = ["logging", "perf-dfa-full"] }
 # So... we don't actually use the regex crate directly. We use the 'meta'
 # regex API from regex-automata. See the README for justification. When
 # bumping the regex crate version, this should be bumped if needed as well.
-regex-automata = "=0.4.3"
+regex-automata = "=0.4.8"
 
 [dependencies.klv]
 path = "../../../shared/klv"