… benchmarks/definitions/curated/01-literal.toml
… benchmarks/definitions/curated/02-literal-alternate.toml
… benchmarks/definitions/curated/07-unicode-character-data.toml
… benchmarks/definitions/curated/09-aws-keys.toml
… benchmarks/definitions/curated/14-quadratic.toml
… benchmarks/definitions/test/func.toml
… benchmarks/definitions/test/model.toml
… benchmarks/haystacks/.gitignore
… engines/hyperscan/Cargo.lock
… engines/icu/Cargo.lock
… engines/pcre2/Cargo.lock
… engines/re2/Cargo.lock
… engines/regress/Cargo.lock
… engines/rust/aho-corasick/Cargo.lock
… engines/rust/memchr/Cargo.lock
… engines/rust/regex-automata/Cargo.lock
… engines/rust/regex-lite/Cargo.lock
… engines/rust/regex-old/Cargo.lock
… engines/rust/regex-syntax/Cargo.lock
… engines/rust/regex/Cargo.lock
… results-2025-02-27.csv