```
$ ./target/release/rebar cmp --intersection results-2025-02-27.csv 
[2025-02-27T12:23:37Z WARN  rebar::format::measurement] curated/09-aws-keys/full:lean: skipping because of error: timeout: exceeded 60s
benchmark                                      lean                     rust/regex           rust/regex/lite
---------                                      ----                     ----------           ---------------
curated/01-literal/sherlock-en                 3.8 MB/s (5002.91x)      18.8 GB/s (1.00x)    24.7 MB/s (777.33x)
curated/01-literal/sherlock-ru                 6.5 MB/s (3079.33x)      19.6 GB/s (1.00x)    51.6 MB/s (389.40x)
curated/01-literal/sherlock-zh                 9.3 MB/s (2806.83x)      25.4 GB/s (1.00x)    69.2 MB/s (375.29x)
curated/02-literal-alternate/sherlock-en       1192.3 KB/s (3732.33x)   4.2 GB/s (1.00x)     12.9 MB/s (336.04x)
curated/02-literal-alternate/sherlock-ru       2029.5 KB/s (1512.90x)   2.9 GB/s (1.00x)     20.0 MB/s (150.10x)
curated/02-literal-alternate/sherlock-zh       2.8 MB/s (1739.89x)      4.7 GB/s (1.00x)     29.4 MB/s (164.56x)
curated/03-date/ascii                          25.7 KB/s (2533.50x)     63.6 MB/s (1.00x)    395.6 KB/s (164.64x)
curated/06-cloud-flare-redos/simplified-short  1052.4 KB/s (788.75x)    810.6 MB/s (1.00x)   6.9 MB/s (118.17x)
curated/06-cloud-flare-redos/simplified-long   1057.0 KB/s (36812.75x)  37.1 GB/s (1.00x)    6.9 MB/s (5498.01x)
curated/07-unicode-character-data/parse-line   1092.9 KB/s (165.06x)    176.2 MB/s (1.00x)   10.9 MB/s (16.22x)
curated/07-unicode-character-data/compile      32.70us (3.31x)          62.12us (6.29x)      9.87us (1.00x)
curated/09-aws-keys/quick                      1248.6 KB/s (730.75x)    891.0 MB/s (1.00x)   12.8 MB/s (69.83x)
curated/09-aws-keys/compile-quick              14.19us (4.03x)          37.86us (10.76x)     3.52us (1.00x)
curated/14-quadratic/1x                        57.8 KB/s (162.97x)      9.2 MB/s (1.00x)     477.4 KB/s (19.72x)
curated/14-quadratic/2x                        29.9 KB/s (177.28x)      5.2 MB/s (1.00x)     241.3 KB/s (21.94x)
curated/14-quadratic/10x                       5.9 KB/s (191.28x)       1132.2 KB/s (1.00x)  51.0 KB/s (22.20x)
$ ./target/release/rebar rank results-2025-02-27.csv 
[2025-02-27T12:24:00Z WARN  rebar::format::measurement] curated/09-aws-keys/full:lean: skipping because of error: timeout: exceeded 60s
Engine           Version           Geometric mean of speed ratios  Benchmark count
------           -------           ------------------------------  ---------------
rust/regex       1.11.0            1.45                            53
rust/regex/lite  0.1.6             25.22                           38
lean             v4.15.0-ae0c91a2  651.76                          17
```